added rbac

+++ frontend/controllers/ParticipController.php
@@ -10,19 +10,13 @@ use yii\rest\Controller;
 use yii\filters\auth\HttpBearerAuth;
 use app\models\Particip;
 use app\models\ParticipSearch;
+use yii\web\ForbiddenHttpException;
 
+class ParticipController extends Controller {
 
-
-/**
- * participController implements the CRUD actions for particip model.
- */
-class ParticipController extends Controller
-{
-    
     /**
      * @inheritdoc
      */
-    
     public function behaviors() {
         $behaviors = parent::behaviors();
         $behaviors['authenticator'] = [
@@ -34,53 +28,33 @@ class ParticipController extends Controller
                 'application/json' => Response::FORMAT_JSON,
             ],
         ];
-        $behaviors['access'] = [
-            'class' => AccessControl::className(),
-                'only' => ['index', 'login', 'logout', 'signup', 'delete'],
-                'rules' => [
-                    [
-                        'allow' => true,
-                        'actions' => ['login', 'signup'],
-                        'roles' => ['?'],
-                    ],
-                    [
-                        'allow' => true,
-                        'actions' => ['logout', 'index', 'delete'],
-                        'roles' => ['@'],
-                    ],
-                ],
-        ];
         return $behaviors;
     }
 
-    /**
-     * Lists all particip models.
-     * @return mixed
-     */
-    /*public function actionOptions()
-    {
-        Yii::$app->response->headers->set('Access-Control-Allow-Origin', '*');
-    } */  
-    
-    public function actionIndex()
-    {
-        $model = new Particip();
-        $data = $model->find()->asArray()->all();
-        return $data;
+    public function actionIndex() {
+        return $this->actionView();
     }
 
-    public function actionView($id = null)
-    {
-        return $this->findModel($id);
+    public function actionView($id = null) {
+        if ($id == null) {
+            if ( \Yii::$app->user->can('admin') ) {
+                $model = new Particip();
+                return $model->find()->all();
+            } else {
+                $model_search = new ParticipSearch();
+                $dataProvider = $model_search->searchCreated();
+                return $dataProvider->query->all(); 
+            }
+        } else {
+            $model = $this->findModel($id);
+            if (!\Yii::$app->user->can('view', ['particip' => $model])) {
+                throw new ForbiddenHttpException('Access denied index');
+            }
+            return $model;
+        }
     }
 
-    /**
-     * Creates a new particip model.
-     * If creation is successful, the browser will be redirected to the 'view' page.
-     * @return mixed
-     */
-    public function actionCreate()
-    {
+    public function actionCreate() {
         $model = new Particip();
 
         if ($model->load(Yii::$app->getRequest()->getBodyParams(), '') && $model->save()) {
@@ -91,16 +65,12 @@ class ParticipController extends Controller
         }
     }
 
-    /**
-     * Updates an existing particip model.
-     * If update is successful, the browser will be redirected to the 'view' page.
-     * @param integer $id
-     * @return mixed
-     * @throws NotFoundHttpException if the model cannot be found
-     */
-    public function actionUpdate($id)
-    {
+    public function actionUpdate($id) {
         $model = $this->findModel($id);
+
+        if (!\Yii::$app->user->can('update', ['particip' => $model])) {
+            throw new ForbiddenHttpException('Access denied update');
+        }
 
         if ($model->load(Yii::$app->request->post(), '') && $model->save()) {
             return ['access_token' => Yii::$app->user->identity->getAuthKey()];
@@ -110,30 +80,20 @@ class ParticipController extends Controller
         }
     }
 
-    /**
-     * Deletes an existing particip model.
-     * If deletion is successful, the browser will be redirected to the 'index' page.
-     * @param integer $id
-     * @return mixed
-     * @throws NotFoundHttpException if the model cannot be found
-     */
-    public function actionDelete($id)
-    {
+    public function actionDelete($id) {
+
         $model = $this->findModel($id);
+
+        if (!\Yii::$app->user->can('delete', ['particip' => $model])) {
+            throw new ForbiddenHttpException('Access denied delete');
+        }
+
         $model->delete();
-          
+
         return $model;
     }
 
-    /**
-     * Finds the particip model based on its primary key value.
-     * If the model is not found, a 404 HTTP exception will be thrown.
-     * @param integer $id
-     * @return particip the loaded model
-     * @throws NotFoundHttpException if the model cannot be found
-     */
-    protected function findModel($id)
-    {
+    protected function findModel($id) {
         if (($model = particip::findOne($id)) !== null) {
             return $model;
         }
